Compile GAOL without the conversion warnings GCC (-Wconversion) and Clang (-Wsign-conversion, which its -Wconversion includes) gave, and regenerate the parser with bison 3.5.1, whose code gives none
With -Wall -Wconversion in the three builds, GCC 9.4 warned twice and Clang
18 sixteen times: implicit conversions between signed and unsigned integers,
and a sum of two unsigned short returned as an unsigned short. None changed
a value that GAOL computes; each is now written with the type it means:

- gaol_common.cpp: save_fpu_state is an unsigned short, what Init_Lib()
  returns and Exit_Lib() takes;
- gaol_interval.cpp, gaol_interval2f.cpp, gaol_intervalf.cpp: pow(I, n) calls
  uipow(I, unsigned) with n and -n converted explicitly, 0u - n for -n,
  which is right for the most negative int too;
- gaol_interval.cpp: the quadrant flags of the trigonometric functions summed
  as ints then converted;
- gaol_interval_sse.cpp: the bytes of the reverse lookup converted to
  uint32_t before the shifts;
- gaol_expression.cpp and gaol_interval_parser.ypp: pow_node takes an int
  and nth_root(), nth_root_node an unsigned int; the casts named the other
  type.

The committed parser (gaol_interval_parser.cpp and .h) was generated by
bison 3.0.4, whose code warned in the skeleton (yytype_int16, the size of
the stacks); regenerated as the Makefile does (bison -y -d -p gaol_) with
bison 3.5.1, whose code compiles without warnings with GCC and Clang. meson,
which regenerates the parser itself, gives the same.

Checked with GCC 9.4 and Clang 18 (CMake): GAOL compiles without any
warning, also with -Wsign-conversion -Wimplicit-int-conversion added for
Clang, and the tests pass; autotools and meson give no warning either.

+++ gaol/gaol_expression.cpp
@@ -1204,7 +1204,7 @@ namespace gaol {
 
   const expression pow(const expression& e, unsigned int n)
   {
-    return *(new pow_node(e,n));
+    return *(new pow_node(e,static_cast<int>(n)));
   }
 
   const expression pow(const expression& e1, const expression &e2)
